Convert between number of DCT bases and Hz of highpass filter

Provide either n or f to calculate the other.

dct_convert(T_, TR, n = NULL, f = NULL)

Arguments

T_
Length of timeseries (number of timepoints)

TR
TR of the fMRI scan, in seconds (the time between timepoints)

n
Number of cosine bases

f
Hz of highpass filter

Value

If n was provided, the highpass filter cutoff (Hz) is returned. Otherwise, if f was provided, the number of cosine bases is returned. The result should be rounded before passing to dct_bases

Details

If only the total length of the scan is known, you can set that to TR and use T_=1.

\(f = n / (2 * T_ * TR)\)