Which components have high kurtosis?

The kurtosis cutoff is a high quantile (default 0.99) of the sampling distribution of kurtosis for Normal iid data of the same length as the components; it is estimated by simulation or calculated from the theoretical asymptotic distribution if the components are long enough.

high_kurtosis(Comps, kurt_quantile = 0.99, n_sim = 5000, min_1 = FALSE)

Arguments

Comps
A matrix; each column is a component. For PCA, this is the U matrix. For ICA, this is the M matrix.

kurt_quantile
components with kurtosis of at least this quantile are kept.

n_sim
The number of simulation data to use for estimating the sampling distribution of kurtosis. Only used if a new simulation is performed. (If \(n<1000\) and the quantile is 90%, a pre-computed value is used instead. If \(n>1000\), the theoretical asymptotic distribution is used instead.)

min_1
Require at least one component to be selected? In other words, if no components meet the quantile cutoff, should the component with the highest kurtosis be returned? Default: FALSE.

Value

A logical vector indicating whether each component has high kurtosis.

Details

The components should not have any strong low-frequency trends, because trends can affect kurtosis in unpredictable ways unrelated to outlier presence.