Robust linear model on DCT bases

Fit a linear model regressing an input vector on DCT bases, robustly.

rob_trend(x, nDCT = 4, lmrob_method = "MM", seed = 0)

Arguments

x
The input vector to regress on DCT bases

nDCT
The number of DCT bases to use. Default: 4

lmrob_method
The lmrob_method argument to robustbase::lmrob.

seed
Set a seed right before the call to robustbase::lmrob? If NULL, do not set a seed. If numeric (default: 0), will use as the seed.

Value

The output of robustbase::lmrob